+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/SpellShopAction.java
@@ -10,6 +10,7 @@ import com.elmakers.mine.bukkit.api.spell.SpellResult;
 import com.elmakers.mine.bukkit.api.spell.SpellTemplate;
 import com.elmakers.mine.bukkit.api.wand.Wand;
 import com.elmakers.mine.bukkit.api.wand.WandUpgradePath;
+import com.elmakers.mine.bukkit.block.MaterialAndData;
 import com.elmakers.mine.bukkit.integration.VaultController;
 import com.elmakers.mine.bukkit.magic.MagicPlugin;
 import com.elmakers.mine.bukkit.spell.BaseSpell;
@@ -26,7 +27,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 public class SpellShopAction extends BaseSpellAction implements GUIAction
 {
@@ -39,6 +43,22 @@ public class SpellShopAction extends BaseSpellAction implements GUIAction
     private boolean useXP = false;
     private CastContext context;
     private Wand wand;
+    private Map<String, Double> spells = new HashMap<String, Double>();
+
+    @Override
+    public void initialize(ConfigurationSection parameters)
+    {
+        super.initialize(parameters);
+        spells.clear();
+        if (parameters.contains("spells"))
+        {
+            ConfigurationSection spellSection = parameters.getConfigurationSection("spells");
+            Collection<String> spellKeys = spellSection.getKeys(false);
+            for (String spellKey : spellKeys) {
+                spells.put(spellKey, spellSection.getDouble(spellKey));
+            }
+        }
+    }
 
     @Override
     public void deactivated() {
@@ -56,33 +76,49 @@ public class SpellShopAction extends BaseSpellAction implements GUIAction
             String spellKey = com.elmakers.mine.bukkit.wand.Wand.getSpell(item);
             SpellTemplate template = context.getController().getSpellTemplate(spellKey);
             boolean isXP = useXP || !VaultController.hasEconomy();
-            double worth = template.getWorth();
+            Double worth = spells.get(spellKey);
+            worth = worth == null ? template.getWorth() : worth;
             boolean hasCosts = false;
             if (worth > 0) {
                 if (isXP) {
-                    hasCosts = mage.getExperience() > (int)worth;
+                    hasCosts = mage.getExperience() > (int)(double)worth;
                 } else {
                     hasCosts = VaultController.getInstance().has(mage.getPlayer(), worth);
                 }
             }
 
             if (!hasCosts) {
-                String costString = isXP ? context.getMessage("insufficient_xp") : context.getMessage("insufficient_money");
-                String costs = isXP ? costString.replace("$cost", Integer.toString((int)worth))
-                        : costString.replace("$cost", VaultController.getInstance().format(worth));
-                context.sendMessage(costs);
-            } else {
-                String costString = isXP ? context.getMessage("deducted_xp") : context.getMessage("deducted_money");
-                String costs = isXP ? costString.replace("$cost", Integer.toString((int)worth))
-                        : costString.replace("$cost", VaultController.getInstance().format(worth));
-                costs = costs.replace("$spell", template.getName());
-                context.sendMessage(costs);
+                String costString = context.getMessage("insufficient_resources");
                 if (isXP) {
-                    mage.removeExperience((int)worth);
+                    String xpAmount = Integer.toString((int)(double)worth);
+                    xpAmount = context.getMessage("costs.xp_amount").replace("$amount", xpAmount);
+                    costString = costString.replace("$cost", xpAmount);
+                } else {
+                    costString = costString.replace("$cost", VaultController.getInstance().format(worth));
+                }
+                context.sendMessage(costString);
+            } else {
+                String costString = context.getMessage("deducted");
+                if (isXP) {
+                    String xpAmount = Integer.toString((int)(double)worth);
+                    xpAmount = context.getMessage("costs.xp_amount").replace("$amount", xpAmount);
+                    costString = costString.replace("$cost", xpAmount);
+                } else {
+                    costString = costString.replace("$cost", VaultController.getInstance().format(worth));
+                }
+                costString = costString.replace("$spell", template.getName());
+                context.sendMessage(costString);
+                if (isXP) {
+                    mage.removeExperience((int)(double)worth);
                 } else {
                     VaultController.getInstance().withdrawPlayer(mage.getPlayer(), worth);
                 }
                 wand.addSpell(spellKey);
+                com.elmakers.mine.bukkit.api.wand.WandUpgradePath path = wand.getPath();
+                WandUpgradePath nextPath = path != null ? path.getUpgrade(): null;
+                if (nextPath != null && autoUpgrade && path.checkUpgradeRequirements(wand, null)) {
+                    path.upgrade(wand, mage);
+                }
             }
             mage.deactivateGUI();
         }
@@ -115,16 +151,20 @@ public class SpellShopAction extends BaseSpellAction implements GUIAction
             context.sendMessage("no_wand");
             return SpellResult.FAIL;
         }
+
         WandUpgradePath path = wand.getPath();
-        if (path == null) {
-            context.sendMessage(context.getMessage("no_upgrade").replace("$wand", wand.getName()));
-            return SpellResult.FAIL;
-        }
+
+        // Check path requirements
         if (requiredPath != null || exactPath != null) {
+            if (path == null) {
+                context.sendMessage(context.getMessage("no_path").replace("$wand", wand.getName()));
+                return SpellResult.FAIL;
+            }
+
             if ((requiredPath != null && !path.hasPath(requiredPath)) || (exactPath != null && !exactPath.equals(path.getKey()))) {
                 WandUpgradePath requiresPath = com.elmakers.mine.bukkit.wand.WandUpgradePath.getPath(requiredPath);
                 if (requiresPath != null) {
-                    context.sendMessage(context.getMessage("no_path").replace("$path", requiresPath.getName()));
+                    context.sendMessage(context.getMessage("no_required_path").replace("$path", requiresPath.getName()));
                 } else {
                     context.getLogger().warning("Invalid path specified in AddSpell action: " + requiredPath);
                 }
@@ -143,27 +183,58 @@ public class SpellShopAction extends BaseSpellAction implements GUIAction
             }
         }
 
-        List<ItemStack> spellItems = new ArrayList<ItemStack>();
-        List<String> availableSpells = new ArrayList<String>(path.getSpells());
-        if (showRequired) {
-            availableSpells.addAll(path.getRequiredSpells());
+        // Load spells
+        Map<String, Double> availableSpells = new TreeMap<String, Double>();
+        if (spells.size() > 0)
+        {
+            availableSpells.putAll(spells);
         }
-        Collections.sort(availableSpells);
+        else
+        {
+            if (path == null) {
+                context.sendMessage(context.getMessage("no_path").replace("$wand", wand.getName()));
+                return SpellResult.FAIL;
+            }
+
+            Collection<String> pathSpells = path.getSpells();
+            for (String pathSpell : pathSpells) {
+                availableSpells.put(pathSpell, null);
+            }
+            if (showRequired) {
+                Collection<String> requiredSpells = path.getRequiredSpells();
+                for (String requiredSpell : requiredSpells) {
+                    availableSpells.put(requiredSpell, null);
+                }
+            }
+        }
+
+        List<ItemStack> spellItems = new ArrayList<ItemStack>();
         MagicAPI api = MagicPlugin.getAPI();
         boolean isXP = useXP || !VaultController.hasEconomy();
-        String costString = isXP ? context.getMessage("cost_xp_lore") : context.getMessage("cost_lore");
-        for (String spellKey : availableSpells) {
+        String costString = context.getMessage("cost_lore");
+        for (Map.Entry<String, Double> spellValue : availableSpells.entrySet()) {
+            String spellKey = spellValue.getKey();
             if (wand.hasSpell(spellKey)) continue;
 
             SpellTemplate spell = api.getSpellTemplate(spellKey);
-            double worth = spell.getWorth();
+            Double worth = spellValue.getValue();
+            if (worth == null) {
+                worth = spell.getWorth();
+                spells.put(spellKey, worth);
+            }
             if (worth <= 0 && !showFree) continue;
 
             ItemStack spellItem = api.createSpellItem(spellKey);
             ItemMeta meta = spellItem.getItemMeta();
             List<String> lore = meta.getLore();
-            String costs = isXP ? costString.replace("$cost", Integer.toString((int)worth))
-                    : costString.replace("$cost", VaultController.getInstance().format(worth));
+            String costs;
+            if (isXP) {
+                String xpAmount = Integer.toString((int)(double)worth);
+                xpAmount = context.getMessage("costs.xp_amount").replace("$amount", xpAmount);
+                costs = costString.replace("$cost", xpAmount);
+            } else {
+                costs = costString.replace("$cost", VaultController.getInstance().format(worth));
+            }
             lore.add(ChatColor.GOLD + costs);
             meta.setLore(lore);
             spellItem.setItemMeta(meta);
@@ -171,16 +242,20 @@ public class SpellShopAction extends BaseSpellAction implements GUIAction
         }
 
         if (spellItems.size() == 0) {
-            WandUpgradePath nextPath = path.getUpgrade();
-            if (nextPath != null && autoUpgrade) {
-                path.upgrade(wand, mage);
-                return SpellResult.CAST;
-            }
             context.sendMessage("no_spells");
             return SpellResult.FAIL;
         }
 
-        String inventoryTitle = context.getMessage("title", "Spell Shop");
+        String inventoryTitle = context.getMessage("title", "Spells ($balance)");
+        if (isXP) {
+            String xpAmount = Integer.toString(mage.getExperience());
+            xpAmount = context.getMessage("costs.xp_amount").replace("$amount", xpAmount);
+            inventoryTitle = inventoryTitle.replace("$balance", xpAmount);
+        } else {
+            double balance = VaultController.getInstance().getBalance(player);
+            inventoryTitle = inventoryTitle.replace("$balance", VaultController.getInstance().format(balance));
+        }
+
         int invSize = ((spellItems.size() + 9) / 9) * 9;
         Inventory displayInventory = CompatibilityUtils.createInventory(null, invSize, inventoryTitle);
         for (ItemStack item : spellItems)